empty virtuoso now uses start-stop-virtuoso script

+++ empty_db_hard.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
-# as ROOT:
+this_script="$(basename $0)"
+if ! [ -e "$this_script" ]; then
+    echo "ERROR, $0 must be run from its own directory"
+    exit 2
+fi
 
-service virtuoso-opensource-7 stop
+./start-stop-virtuoso stop
+
+this_dir=$(pwd)
 cd /var/lib/virtuoso-opensource-7
 rm -rf db.old
 mv db db.old
 tar xvzf db-empty.tar.gz 
-service virtuoso-opensource-7 start
+cd $this_dir
 
+./start-stop-virtuoso start
